+++ crates/router/src/core/authentication/types.rs
@@ -1,43 +1,13 @@
-use cards::CardNumber;
 use error_stack::{Report, ResultExt};
-use serde::{Deserialize, Serialize};
+pub use hyperswitch_domain_models::router_request_types::authentication::{
+    AcquirerDetails, ExternalThreeDSConnectorMetadata, PreAuthenticationData, ThreeDsMethodData,
+};
 
 use crate::{
-    core::{errors, payments},
+    core::errors,
     types::{storage, transformers::ForeignTryFrom},
     utils::OptionExt,
 };
-pub enum PreAuthenthenticationFlowInput<'a, F: Clone> {
-    PaymentAuthNFlow {
-        payment_data: &'a mut payments::PaymentData<F>,
-        should_continue_confirm_transaction: &'a mut bool,
-        card_number: CardNumber,
-    },
-    PaymentMethodAuthNFlow {
-        card_number: CardNumber,
-        other_fields: String, //should be expanded when implementation begins
-    },
-}
-
-pub enum PostAuthenthenticationFlowInput<'a, F: Clone> {
-    PaymentAuthNFlow {
-        payment_data: &'a mut payments::PaymentData<F>,
-        authentication: storage::Authentication,
-        should_continue_confirm_transaction: &'a mut bool,
-    },
-    PaymentMethodAuthNFlow {
-        other_fields: String, //should be expanded when implementation begins
-    },
-}
-
-#[derive(Clone, Debug)]
-pub struct PreAuthenticationData {
-    pub threeds_server_transaction_id: String,
-    pub message_version: common_utils::types::SemanticVersion,
-    pub acquirer_bin: Option<String>,
-    pub acquirer_merchant_id: Option<String>,
-    pub connector_metadata: Option<serde_json::Value>,
-}
 
 impl ForeignTryFrom<&storage::Authentication> for PreAuthenticationData {
     type Error = Report<errors::ApiErrorResponse>;
@@ -61,21 +31,4 @@ impl ForeignTryFrom<&storage::Authentication> for PreAuthenticationData {
             connector_metadata: authentication.connector_metadata.clone(),
         })
     }
-}
-
-#[derive(Clone, Default, Debug, Serialize, Deserialize)]
-pub struct ThreeDsMethodData {
-    pub three_ds_method_data_submission: bool,
-    pub three_ds_method_data: String,
-    pub three_ds_method_url: Option<String>,
-}
-#[derive(Clone, Default, Debug, Serialize, Deserialize)]
-pub struct AcquirerDetails {
-    pub acquirer_bin: String,
-    pub acquirer_merchant_id: String,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-pub struct ExternalThreeDSConnectorMetadata {
-    pub pull_mechanism_for_external_3ds_enabled: Option<bool>,
 }
