+++ crates/router/src/core/payments/types.rs
@@ -9,7 +9,9 @@ use common_utils::{
 use diesel_models::business_profile::BusinessProfile;
 use error_stack::ResultExt;
 use hyperswitch_domain_models::payments::payment_attempt::PaymentAttempt;
-pub use hyperswitch_domain_models::router_request_types::{AuthenticationData, SurchargeDetails};
+pub use hyperswitch_domain_models::router_request_types::{
+    AuthenticationData, PaymentCharges, SurchargeDetails,
+};
 use redis_interface::errors::RedisError;
 use router_env::{instrument, tracing};
 
@@ -373,11 +375,4 @@ impl ForeignTryFrom<&storage::Authentication> for AuthenticationData {
             Err(errors::ApiErrorResponse::PaymentAuthenticationFailed { data: None }.into())
         }
     }
-}
-
-#[derive(Debug, serde::Deserialize, Clone)]
-pub struct PaymentCharges {
-    pub charge_type: api_models::enums::PaymentChargeType,
-    pub fees: i64,
-    pub transfer_account_id: String,
 }
