+++ crates/router/src/routes/fraud_check.rs
@@ -1,11 +1,9 @@
 use actix_web::{web, HttpRequest, HttpResponse};
-use common_utils::events::{ApiEventMetric, ApiEventsType};
 use router_env::Flow;
 
 use crate::{
     core::{api_locking, fraud_check as frm_core},
     services::{self, api},
-    types::fraud_check::FraudCheckResponseData,
     AppState,
 };
 
@@ -27,10 +25,4 @@ pub async fn frm_fulfillment(
         api_locking::LockAction::NotApplicable,
     ))
     .await
-}
-
-impl ApiEventMetric for FraudCheckResponseData {
-    fn get_api_event_type(&self) -> Option<ApiEventsType> {
-        Some(ApiEventsType::FraudCheck)
-    }
 }
